requirements.txt
pandas
numpy
psycopg2-binary
faker
python-dotenv
scikit-learn
matplotlib
reportlab
